websites/roteiro-petropolis-website/src/agent/requirements.txt
browser-use
pandas
schedule
python-dotenv


